requirements.txt
streamlit==1.28.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.12.0
chromadb==0.3.29
# pysqlite3
# pandas>=2.1.0
# numpy>=1.25.0
# scikit-learn>=1.3.0
# joblib>=1.3.2
python-dotenv>=1.0.0
tavily-python>=0.3.0
faiss-cpu>=1.7.4
pydantic>=2.5.0
# xgboost
# shap
langchain_deepseek